greedy: FCFS; shorestDistance
